Fixed backend + some refactoring

+++ LegoManager-api/src/main/java/cz/muni/fi/pa165/legomanager/services/impl/LegoKitServiceImpl.java
@@ -5,7 +5,6 @@
  */
 package cz.muni.fi.pa165.legomanager.services.impl;
 
-import cz.muni.fi.pa165.legomanager.dao.LegoDaoException;
 import cz.muni.fi.pa165.legomanager.dao.LegoKitDao;
 import cz.muni.fi.pa165.legomanager.entity.LegoKit;
 import cz.muni.fi.pa165.legomanager.services.LegoKitService;
@@ -13,9 +12,7 @@ import cz.muni.fi.pa165.legomanager.transferobjects.LegoKitTO;
 import java.util.ArrayList;
 import java.util.List;
 import org.dozer.DozerBeanMapper;
-import org.dozer.MappingException;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.dao.RecoverableDataAccessException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -29,6 +26,7 @@ public class LegoKitServiceImpl implements LegoKitService {
 
     @Autowired
     LegoKitDao legoKitDao;
+    
     @Autowired
     DozerBeanMapper mapper;
 
@@ -36,56 +34,37 @@ public class LegoKitServiceImpl implements LegoKitService {
     public List<LegoKitTO> getAllLegoKits() {
         
         List<LegoKitTO> legoKitsTO = new ArrayList<>();
-        try {
-            for (LegoKit kit : legoKitDao.getAllLegoKits()) {
-                legoKitsTO.add(mapper.map(kit, LegoKitTO.class));
-            }
-        } catch (IllegalArgumentException | LegoDaoException | MappingException ex) {
-            throw new RecoverableDataAccessException("error while retrieving data", ex);
-        }
+        for (LegoKit kit : legoKitDao.getAllLegoKits()) {
+            legoKitsTO.add(mapper.map(kit, LegoKitTO.class));
+        }       
         return legoKitsTO;
     }
 
     @Override
-    public void updateLegoKit(LegoKitTO legoKit) {
+    public void updateLegoKit(LegoKitTO legoKit) {      
+        if(legoKit==null) throw new IllegalArgumentException();
+        legoKitDao.updateLegoKit(mapper.map(legoKit, LegoKit.class));
         
-        try {
-            legoKitDao.updateLegoKit(mapper.map(legoKit, LegoKit.class));
-        } catch (MappingException | IllegalArgumentException | LegoDaoException ex) {
-            throw new RecoverableDataAccessException("error while updating lego kit", ex);
-        }
     }
 
     @Override
     public void deleteLegoKit(LegoKitTO legoKit) {
-
-        try {
-            legoKitDao.deleteLegoKit(mapper.map(legoKit, LegoKit.class));
-        } catch (MappingException | IllegalArgumentException | LegoDaoException ex) {
-            throw new RecoverableDataAccessException("error while deleting lego kit", ex);
-        }
+        if(legoKit==null) throw new IllegalArgumentException();
+        legoKitDao.deleteLegoKit(mapper.map(legoKit, LegoKit.class));
+       
     }
 
     @Override
-    public void createLegoKit(LegoKitTO legoKit) {
-
-        try {
-            legoKitDao.addLegoKit(mapper.map(legoKit, LegoKit.class));
-        } catch (MappingException | IllegalArgumentException | LegoDaoException ex) {
-            throw new RecoverableDataAccessException("error while creating lego kit", ex);
-        }
+    public void createLegoKit(LegoKitTO legoKit) {        
+        if(legoKit==null) throw new IllegalArgumentException();
+        legoKitDao.addLegoKit(mapper.map(legoKit, LegoKit.class));
+       
     }
 
     @Override
     public LegoKitTO getLegoKit(Long id) {
-        
-        LegoKitTO legoKit;
-        try {
-            legoKit = mapper.map(legoKitDao.findLegoKitById(id), LegoKitTO.class);
-        } catch (IllegalArgumentException | LegoDaoException | MappingException ex) {
-            throw new RecoverableDataAccessException("error while retrieving lego kit", ex);
-        }
-        return legoKit;
+        if(id==null) throw new IllegalArgumentException();
+        return mapper.map(legoKitDao.findLegoKitById(id), LegoKitTO.class);       
     }
 
     public void setLegoKitDao(LegoKitDao legoKitDao) {
