+++ LegoManager-api/src/main/java/cz/muni/fi/pa165/legomanager/services/impl/LegoPieceServiceImpl.java
@@ -5,7 +5,6 @@
  */
 package cz.muni.fi.pa165.legomanager.services.impl;
 
-import cz.muni.fi.pa165.legomanager.dao.LegoDaoException;
 import cz.muni.fi.pa165.legomanager.services.LegoPieceService;
 import cz.muni.fi.pa165.legomanager.transferobjects.LegoPieceTO;
 import java.util.List;
@@ -16,9 +15,6 @@ import cz.muni.fi.pa165.legomanager.dao.LegoPieceDao;
 import cz.muni.fi.pa165.legomanager.entity.LegoPiece;
 import java.util.ArrayList;
 import org.dozer.DozerBeanMapper;
-import org.dozer.MappingException;
-import org.springframework.dao.DataAccessException;
-import org.springframework.dao.RecoverableDataAccessException;
 
 /**
  *
@@ -41,63 +37,44 @@ public class LegoPieceServiceImpl implements LegoPieceService{
     public void setMapper(DozerBeanMapper mapper) {
         this.mapper = mapper;
     }
-    
-    
-    
+      
     @Override
     @Transactional
-    public void createLegoPiece(LegoPieceTO legoPiece) throws DataAccessException {
-        try{
-            legoPieceDao.addLegoPiece(mapper.map(legoPiece, LegoPiece.class));
-        }catch(IllegalArgumentException | LegoDaoException | MappingException ex){
-            throw new RecoverableDataAccessException("Error while creating lego piece", ex);
-        }
+    public void createLegoPiece(LegoPieceTO legoPiece) {
+        if(legoPiece==null) throw new IllegalArgumentException();
+        legoPieceDao.addLegoPiece(mapper.map(legoPiece, LegoPiece.class));     
     }
 
     @Override
     @Transactional
-    public void updateLegoPiece(LegoPieceTO legoPiece) throws DataAccessException {
-        try{
-            legoPieceDao.updateLegoPiece(mapper.map(legoPiece, LegoPiece.class));
-        }catch(IllegalArgumentException | LegoDaoException | MappingException ex){
-            throw new RecoverableDataAccessException("Error while updating lego piece", ex);
-        }
+    public void updateLegoPiece(LegoPieceTO legoPiece) {
+        if(legoPiece==null) throw new IllegalArgumentException();
+        legoPieceDao.updateLegoPiece(mapper.map(legoPiece, LegoPiece.class));    
     }
 
     @Override
     @Transactional
-    public void removeLegoPiece(LegoPieceTO legoPiece) throws DataAccessException {
-        try{
-            legoPieceDao.deleteLegoPiece(mapper.map(legoPiece, LegoPiece.class));
-        }catch(IllegalArgumentException | LegoDaoException | MappingException ex){
-            throw new RecoverableDataAccessException("Error while deleting lego piece", ex);
-        }
+    public void removeLegoPiece(LegoPieceTO legoPiece) {
+        if(legoPiece==null) throw new IllegalArgumentException();
+        legoPieceDao.deleteLegoPiece(mapper.map(legoPiece, LegoPiece.class));
+        
     }
 
     @Override
     @Transactional
-    public LegoPieceTO getLegoPiece(Long id) throws DataAccessException {
-        LegoPieceTO legoPiece;
-        try{
-            legoPiece = mapper.map(legoPieceDao.findLegoPieceById(id), LegoPieceTO.class);
-        }catch(IllegalArgumentException | LegoDaoException | MappingException ex){
-            throw new RecoverableDataAccessException("Error while retrieving lego piece", ex);
-        }
-        return legoPiece;
+    public LegoPieceTO getLegoPiece(Long id) {
+        if(id==null) throw new IllegalArgumentException();
+        return mapper.map(legoPieceDao.findLegoPieceById(id), LegoPieceTO.class);      
     }
 
     @Override
     @Transactional
-    public List<LegoPieceTO> getAllLegoPieces() throws DataAccessException {
+    public List<LegoPieceTO> getAllLegoPieces() {
         List<LegoPieceTO> legoPieceTOs;
-        try {
-            List<LegoPiece> legoPieces = legoPieceDao.getAllLegoPieces();
-            legoPieceTOs = new ArrayList<>();
-            for (LegoPiece piece : legoPieces) {
-                legoPieceTOs.add(mapper.map(piece, LegoPieceTO.class));
-            }
-        } catch (IllegalArgumentException | LegoDaoException | MappingException ex) {
-            throw new RecoverableDataAccessException("Error while retrieving data", ex);
+        List<LegoPiece> legoPieces = legoPieceDao.getAllLegoPieces();
+        legoPieceTOs = new ArrayList<>();
+        for (LegoPiece piece : legoPieces) {
+            legoPieceTOs.add(mapper.map(piece, LegoPieceTO.class));
         }
         return legoPieceTOs;
     }
